feat: Changelog constants extraño
